requirements.txt
django==2.2.5
django-rest-framework==0.1.0
